feat: 컴포넌트 분할 및 이벤트 버블링 처리 App에서 Header로 Component 분할 Component.js에 addEvent 메소드를 생성해서 이벤트 버블링 처리
