+++ src/views/components/common/Component.js
@@ -1,27 +1,52 @@
 export default class Component {
-	$target;
-	props;
-	state;
-	constructor($target, props) {
-		this.$target = $target;
-		this.props = props;
-		this.setup();
-		this.setEvent();
-		this.render();
-	}
-	setup() {}
-	mounted() {}
-	template() {
-		return ``;
-	}
-	render() {
-		this.$target.innerHTML = this.template();
-		this.mounted();
-	}
-	setEvent() {}
-	setState(newState) {
-		this.state = { ...this.state, ...newState };
-		this.render();
-	}
-	addEvent(eventType, selector, callback) {}
+  $target;
+  props;
+  state;
+
+  constructor($target, props) {
+    this.$target = $target;
+    this.props = props;
+    this.setup();
+    this.setEvent();
+    this.render();
+  }
+
+  setup() {}
+  mounted() {}
+  template() {
+    return ``;
+  }
+
+  render() {
+    this.$target.innerHTML = this.template();
+    this.mounted();
+  }
+
+  setEvent() {}
+  setState(newState) {
+    this.state = { ...this.state, ...newState };
+    this.render();
+  }
+  addEvent(eventType, selector, callback) {
+    this.$target.addEventListener(eventType, (event) => {
+      if (!event.target.closest(selector)) return false;
+      callback(event);
+    });
+  }
+  removeEvent(eventType, selector, callback) {
+    this.$target.removeEventListener(eventType, (event) => {
+      if (!event.target.closest(selector)) return false;
+      callback(event);
+    });
+  }
+  addWindowEvent(eventType, callback) {
+    window.addEventListener(eventType, (event) => {
+      callback(event);
+    });
+  }
+  removeWindowEvent(eventType, callback) {
+    window.removeEventListener(eventType, (event) => {
+      callback(event);
+    });
+  }
 }
